some renaming + added report service

+++ assignment2/app/src/main/java/at/ac/tuwien/nsa/gr12/comparelocations/adapter/mozilla/location/service/model/MLSResponse.kt
@@ -2,7 +2,7 @@ package at.ac.tuwien.nsa.gr12.comparelocations.adapter.mozilla.location.service.
 
 class MLSResponse(
     var location: MLSLocation? = null,
-    var accuracy: Double? = null
+    var accuracy: Float? = null
 ) {
     constructor() : this(null, null)
 }
