+++ assignment2/app/src/main/java/at/ac/tuwien/nsa/gr12/comparelocations/core/model/Location.kt
@@ -3,7 +3,7 @@ package at.ac.tuwien.nsa.gr12.comparelocations.core.model
 data class Location(
     var latitude: Double? = null,
     var longitude: Double? = null,
-    var accuracy: Int? = null
+    var accuracy: Float? = null
 ) {
     constructor() : this(null, null, null)
 }
